Fix C=256 OOB crash + slot_mapping type mismatch + upgrade build to CUDA 13.2
Bug 1: C++ store kernel expected int32 slot_mapping but vLLM passes int64.
Fix: cast slot_mapping.int() before passing to torch.ops.fusencache.store_kv.

Bug 2: Shared decode buffers sized with _max_B = cg_max which REPLACED
max_num_seqs instead of taking the max. If cg_max < max_num_seqs or was
read before finalization, buffers would be under-allocated → OOB at C=256.
Fix: _max_B = max(_max_B, cg_max).

Also updated docker/build_vllm.sh for CUDA 13.2 + latest vLLM main.

+++ kv_cache_gen/generate.py
@@ -446,7 +446,7 @@ def make_store_fn(spec: KVCacheSpec):
                 max_slots, Hk, num_sb, 2, dtype=torch.float16, device=key.device)
 
         _cpp_store_op(
-            k, v, kv_cache, layer._fc_scales, slot_mapping,
+            k, v, kv_cache, layer._fc_scales, slot_mapping.int(),
             D, block_size,
             spec.k_bits, spec.v_bits,
             spec.k_scale_block, spec.v_scale_block,
